Refactor Job starting from RPC to C2.
With the ability to persistently start jobs, it no longer makes
sense for the code that actually stops the jobs to be in the rpc
package.  C2 seemed to be optimal place to move them, since they
listen for incoming C2 connections.
